requirements.txt
asgiref==3.9.2
boto3==1.40.40
botocore==1.40.40
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.4.0
django-redis==5.4.0
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
stripe==12.5.1
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
